requirements.txt
python-telegram-bot[job-queue]==20.3
python-dotenv==1.0.0
sqlalchemy==2.0.0
alembic==1.12.0
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.0.0
isort==5.12.0
flask==2.3.0
gunicorn==20.1.0